58 handle missing docs (#60)
* Handle sites that support 2FA and do not have docs

If the site supports 2FA and does not have docs,
then we tell the user that the site supports 2FA
and tell them to contact the company for details.

Does it suck? Yes. Not many ideas on how to handle
this right now. Definite candidate for improvement
in the future.

See #58 for details.

* Updates CHANGELOG