Add entry for www.google.com
Big change to the data schema here. I made the img
field optional instead of required.

At first, I did this because I didn't have a Google logo
image and wanted to just move quickly since 2FAN does not
currently use images for anything (and we don't have plans
to start using them for something right now).

However, I then realized that all of the images are hosted on
twofactorauth.org and I would have to wait for a PR to go
through their process. I do not think that it is a good idea
to rely on image URLs that are controlled by another project
for now because it limits our ability to move quickly.

We can come back to clean up data and add in logos in the
future if we decide that we need/want them and also have a
solution for where to host them. It should be trivial to
host them on 2fanotifier.org if we want to do that in the
future.