Initial on-boarding tutorial (#74)
* first attempt

* layout done

layout is at a good point. Need to add in animation for the elements to point out. And also need to add some accessibility elements.

* animations created

Created simple animations for helping focus user's attention. Could definitely be done better with JS, but this works for now until we can get more feedback.

* Tweak tutorial wording

I made these changes in real time while we were on a retro call.

* updated changelog