Corrects handling of tabs.onUpdated event (#94)
* Corrects handling of tabs.onUpdated event

I was processing the current tab each time the
tabs.onUpdated event fired, which was incorrect
and caused a bunch of noise. That event is really
noisy and fires each time anything updates in a
Tab. However, I only want to process the tab once
it is finished loading, so I now check to make sure:
1) The event applies to the current Tab. This is
   useful to prevent events from Tabs which are
   not the current Tab. This can happen if the user
   starts to load a page and then quickly jumps to
   other tabs before the page finishes loading, etc.
2) The event has the status of 'complete', which
   means that the tab has finished loading.

These changes should eliminate duplicate notifications.

* Update change log