89 already enabled feature (#91)
* Revisions made to confirmation page

Created illustration. Made revisions to the popup page and the confirmation page. Created an additional "previously-enabled" page for when users return to a site that they already have flagged as "2fa enabled."

* Updated changelog

* Added images

I'm an idiot and forgot to actually place the images in.

* Fix spelling typo

* Update JS based on proposed HTML changes

* Adjusted text content

Adjusted text for the "already enabled here" page.

* Moves common function into utils.js

userAlreadyEnabled2faOnSite() is going to be used
in multiple places, so I'm moving it into the common
utils.js file and exporting it for use.

* work in progress

added js for "already enabled 2fa" page

* Corrects comment explaining parameter

* Correctly displays previouslyEnabled2FA page

- Fixes bug where we were not waiting for the async
  function call to complete in an if statement. This
  was causing us to constantly show the wrong page
  because the function exists, so the if statement
  resolves to true every time
- Renames page to previouslyEnabled2FA.html
- Updates page to have its own JS file
- Updates JS to update the site name and doc URL

* Fixed images and bolded site name

* Remove unused file

* Addresses choppy navigation in popup pages

It was really common for the popup.html page
to get shown before the JS executed that redirected
the page to supports2FA.html. This meant that the
user could briefly see the "No 2FA here!" page
before being redirected. It was glitchy and weird.

I addressed it temporarily by making the popup.html
page have no visual HTML at all and instead just
have JS. This JS still takes a second to execute
and redirect the page (so, we'll need a better
solution eventually), but at least the user no longer
sees a "No 2FA here!" message and then get redirected
to a "Yipee! 2FA!" page. That was confusing and bad.

* Reword message and bold site name

* Bold site name and format text

* Update changelog